pyproject.toml
[project]
name = "notes-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "ipython>=9.2.0",
]
